Removed isStatic + isAbstract from base member class
