+++ lib/Core/Reflection/ReflectionProperty.php
@@ -7,21 +7,7 @@ use Phpactor\WorseReflection\Core\Visibility;
 use Phpactor\WorseReflection\Core\DocBlock\DocBlock;
 use Phpactor\WorseReflection\Core\Types;
 
-interface ReflectionProperty extends ReflectionNode
+interface ReflectionProperty extends ReflectionMember
 {
-    public function position(): Position;
-
-    public function declaringClass(): ReflectionClassLike;
-
-    public function class(): ReflectionClassLike;
-
-    public function name(): string;
-
-    public function visibility(): Visibility;
-
-    public function inferredTypes(): Types;
-
     public function isStatic(): bool;
-
-    public function docblock(): DocBlock;
 }
